Include date in the default name for the export-all zip file
...and let to the user to change the format of it via the
  extensions.xseei.exportAll.defaultFileName preference. The accepted
format choices are the same as strftime() in C:
  http://pubs.opengroup.org/onlinepubs/007908799/xsh/strftime.html

The character '/', bad as part of a filename in all filesystem but
usually present in date strings, is replaced by a safe alternative.
Other characters forbidden in specific filesystems (e.g. Windows)
aren't specially treated, as the filepicker from the OS will manage
it anyway.