Set defaultExtension attribute for filepickers
It automatically appends the given extension to the filename set by the
user in the export dialogs, if required, on some platforms (MS Windows,
at least, where file associations are strongly depending of it).

Given the above, appending the extension to the default filename offered
to the user is still useful.