Update comments to match re-arrangement of Fx preferences
Nightly 55 saw a re-arrangement of the contents of the different
panes of the Preferences page.

https://bugzilla.mozilla.org/show_bug.cgi?id=1335907

It has not affected to the add-on, so the unique changes done here
are internal comments updates. Accesskeys will be not reseted due to
being almost impossible to find unique accesskeys in the crowded
"General" pane where search engines can be managed now.