Fix files with wrong encoding in export-all zip file
If an installed engine have some property (e.g. name, description)
with non-ascii characters, the respective XML file exported via
the "Export all as a ZIP file" willn't have the UTF-8 file encoding,
still when it's marked as such, and such file can't be imported back
to Firefox as it is.

This issue has been since the initial implementation of the export-all
feature; I should not have missed it to begin with :(  It doesn't happen
when the engine is individually exported.

To read:
https://developer.mozilla.org/en-US/docs/Mozilla/Tech/XPCOM/Guide/Streams#A_note_about_Unicode_strings_versus_nsIInputStream